Revamp Resources page
Up until now the resource page has consisted of mainly links leading to
external sites, however with more internal links being added, we
decided to change the structure from a single file to a module.

The new module's index.js file contains the react component and
information to display all the links on the page. Internal links also
have either a component or a md field depending on whether they are a
react or markdown page - which are imported from submodules of the new
Resources module.