Revamp Destination pages
Just as with the Resource pages, I changed the destination pages from a
react component that would load a folder of markdown pages into a node
module that default exports the react component.

Each page is now self contained in its own directory and can be either
a markdown page or a react component.