Remove description from event cards on mobile
The length of event descriptions varies greatly between events, and the
calendar was having a real estate problem with some of the larger
events.

There is now a new 'ToggleView' component that renders a "See More"
message, and when clicked will become its children.

This component is used to hold the event descriptions so that they dont
show by default, but the user is not directed away from the calendar
when inspecting an event.