Fix calendar event fetch bug
When we display the days in a month, sometimes the month begins after
the start or ends before the end of a week. To make the visual
expreience more complete, we display the days of the bordering months
in a darker colour.

Up until now, if there was an event on a day outside of the month, it
would not show because of a small quirk with the moment library. When
adding or subtracting time from a moment object, the function takes
two parameters, a number and a unit. With no unit, the calculation was
not being performed correctly.

This has now been resolved.